adicionando os exercicios 1 e 2

+++ src/logicadeprogramacao/estrutural/Exercicio1.java
@@ -7,8 +7,8 @@ public class Exercicio1 {
      // Programa para ler a temperatura em graus Celsius e Fahrenheit
      Scanner scanner = new Scanner(System.in);
      System.out.println("Informe a temparatura em graus Celcius");
-     int c = scanner.nextInt();
-     int fahrenheit = (9 * c + 160) /5;
+     double c = scanner.nextDouble();
+     double fahrenheit = (double) (9 * c + 160) /5;
      System.out.println("A temperatura " + c+ " graus celcius convertida em Fahrenheit é " + fahrenheit);
      scanner.close();
